Re-apply change 945e079e60506e88435802d0a764c8b4121ad16f by Richard Moe Gustavsen
BT: JavaScriptCore makes applications crash on leopard
JavaScriptCore checks the Mac OS version when building, and uses
symbols only defined in i e.g.snow leopard when building on snow
leopard.
This means that if you build Qt on snow leopard and later move
it to leopard, applications will crash looking for a missing
symbol (pthread_setname_np in particular).
In Qt, we garuantee that you can run your applications on 10.4 and up
for Carbon, and 10.5 and up for Cocoa. So using compile time checks
this way is not a proper soulution.

Result: In Qt, never use symbols not defined on all supported OS
versions, or at least implement run-time checks.

Rev-By: Simon Hausmann
Rev-By: Tor Arne
Rev-By: MortenS