Switch on DEF files
DEF files containing the frozen DLL exports are in use for Symbian OS
builds, except developer builds (configured with -developer-build)

The reason for this exception is that developer builds export additional
private interfaces to allow autotests to inject or monitor the internal
data of a class.

These autotest exports are not part of the API or the binary interface,
so they are excluded from DEF files.

Task-number: QTBUG-4436
Reviewed-by: Jason Barron