Qt covers BC break in Symbian
Workaround: fntstore.h has an inlined function
'COpenFont* CBitmapFont::OpenFont()' that returns a private data member.
The header will change between minor SDK versions, thus break BC. But
Qt has to build on any SDK version and run on other versions of Symbian
OS. Also Qt does not want to deliver that BC to Qt based apps.

This hack performs the needed pointer arithmetic to get the right
COpenFont* pointer, no matter if the 'Flexible Memory Model' is already
supported or not.

The author is not proud of this commit.

Task-number: QT-2250
Reviewed-by: Iain
Reviewed-by: Shane Kearns

	modified:   src/gui/text/qfontdatabase_s60.cpp