Fix for qmessagebox default softkey visibility + some tab/space fixes.
Default actions get added after the QDialogButtonBox::event receives
show event. That's why 'ok' softkey action will not get added to the
widget. Changed implementation so that softkey actions are added to
widget when they are created.

This implementation is also more safe in a sence that if actions
are changed when message box is visible, the softkeys will be updated.

Task-number: QTBUG-5644
Task-number: QT-2211
Reviewed-by: Sami Merila