Made widgets not clear their contents after losing edit focus.
This is a behavior change from Qt 4.5. The motivation for the change
is the following use case: A user types into a line or text edit, and
gets interrupted by for example phone call or an SMS. This causes the
active application to lose focus, which again causes edit focus to be
lost. When resuming the application the focus gets set back, but the
edit focus does not (nor should it). However, when the user types the
next character, the line edit will erase *all* contents of the widget
because it did not have edit focus. This is extremely destructive for
a mobile phone user, especially since undo is not easily available.

Another reason for doing that fix is that it seems that this is the
mode of operation that most existing phones use.

The fix is enabled for Symbian only at the moment, but will be
followed by a patch for all platforms which will go into 4.7.

Task:     QTBUG-5698
RevBy:    Jason Barron