Improved compatibility with MDI area custom themes on XP
With certain custom Windows themes the window mask will be
incorrectly set so that the subwindows become unusable.
Instead of passing the window rect to the mask region
function, we now only pass only the title bar rect and add
the remainder at the bottom. This should improve
compatibility with certain custom themes such as the
Vista Inspirat Pack.

Task-id: QTBUG-6271
Reviewed-by: prasanth