Add option to ingore unused navigation events in QTextControl
QFxTextEdit, as well as systems with keypad navigation, needs the text
control to ignore navigation events that have no effect (like hitting
the right navigation key at the end of the text.) Previously, we would
only support this for keypad navigation. This patch introduces a flag
that you can set to tell the QTextControl to ignore these events. If
the flag is not explicitly set, behavior should remain as before.
The if-test has been refactored to make it more readable.

Done-with: Alan Alpert
Reviewed-by: Samuel
(cherry picked from commit fab1ce65da8bacfce92b1df7656780e729d31b74)