Cocoa: QColorDialog makes the application hang
The reason is that the color dialog is using a native dialog.
And this dialog enters modality on its own, bypassing the
modality mechanism in Qt. Most important, when the dialog leaves
modality, it does so by calling [NSApp stopModal]. And this call
stops all modal sessions in the application, including the ones
owned by Qt.

This patch detects when this situation occurs, and makes
sure that all modal sessions gets stopped the correct way, and
later rebuilt.

Rev-By: MortenS