Tab shapes have the wrong size and distance between them on Symbian
With certain themes it is really apparent that the tabshape rect and
position deduction in QS60Style has some issues. See for example
5800Xm's default theme. TabShapes have huge spaces between them.

Problem is that the initializer in style for tab shape assumes
that the side parts of the frame are squares and will use 1:1 ratio
(width:height) when upscaling the theme graphic draw area.
This leads to that tabshape graphics will not fill in the
whole area reserved for them.

When initial area has ratio of 1:2 (for horizontal tab shape),
upscaling pruduces correct rect and graphic is drawn correctly.

Task-number: QTBUG-5659
Reviewed-by: Alessandro Portale
(cherry picked from commit 3826d81d96c53f253f2f75683cd1802faa1628df)