Fixes QGraphicsWidget: paint() being called before polish().
The problem is that we request an update() before we schedules a polish
event in QGraphicsScene::addItems, which means paint() is being called before
polishEvent(). We could try to swap the order in addItems, but that
doesn't give us any guarantee that polish is delivered before update
(since we have no control over what's happening from outside graphics
view). A better solution is to always make sure we don't have unpolished
items before we draw.

Auto-test included.

Task-number: QTBUG-4979
Reviewed-by: Andreas