Fixed: QCalendarWidget::showNextMonth() followed by a click on the navigation button causes it to go to the first possible month
QCalendarWidget::setCurrentPage did not set the current index on the
model.  This is required in order to get QCalendarWidgetPrivate::getCurrentDate
working. (and this is used by the calendar widget to navigate)

This intentionaly do not check if the date is inside the minimumDate or
maximumDate range.  This was possible before, and the autotests tests
that behaviour.

Task-number: QTBUG-4058
Reviewed-by: Prasanth Ullattil