Fixed maximized dialogs losing maximized status on orientation switch
If maximized dialog had minimum size that didn't fully fit to
the screen, it lost its maximized status when orientation was switched,
because of the logic that assumed that if a window is resized, it must
no longer be maximized. Skipped this assumption for cases where resize
occurs because enforcement of the minimum size of the window.

Task-number: QTBUG-4671
Reviewed-by: Janne Anttila
Reviewed-by: Sami Merila
(cherry picked from commit ee8e0fc9f32a84bd1e1673c9756044d62ecaebf5)