Fixed stroking of discontinuous paths with the GL2 engine.
Since the stroke is rendered as a triangle strip, zero area triangles
must be inserted in order to have gaps in the stroke. This is achieved
by duplicating vertices before and after each gap. It was already done
for open subpaths. This commit fixes gaps between closed subpaths.

Task-number: QTBUG-5736
Reviewed-by: Gunnar