Character spacing when drawing a QPicture to a high DPI device.
Drawing some text into a QPicture, and then drawing that QPicture
to e.g. a QPrinter, would make the character spacing all wrong.
The old approach that tried to fix this adjusted the DPI of the
font that was used, so that the text size would be correct.
It seems this is not enough, since it will lay out the text
in the default resolution, and then scale that up, which may/may not
introduce errors.

Task-number: QTBUG-4974, 256481
Reviewed-by: Kim