Designer crashes in Mac Cocoa port.
When we reparent a widget, it has to be removed from the old view before
it is added to the new one. Otherwise this can cause the old view to
have an invalid first responder, causing crash.

Task-number: QTBUG-5327
Reviewed-by: Denis
(cherry picked from commit d215ecd857be9eedee03703153ef2509c01b4811)