Define own palette for web-based widgets in QS60Style
It makes no sense to use the S60-theme based palette in web based
widgets. Web pages do not generally show theme background, but their
own (or by default white) which clashes badly with WindowText and Text
that are set to work with theme background. Therefore, QWebView and
QGraphicsWebView need to be set their own palette for texts in
theme palette hash. This improves visibility of text in webkit
when QS60Style is active.

Task-number: QTBUG-4885
Reviewed-by: Janne Koskinen