Remove needless loop in QIODevice::seek
Either the buffer has more data than is being skipped and it can be
cleared or it doesn't and we must skip it.

If the total size of QRingBuffer exceeds INT_MAX there'll be bigger
problems (e.g., QRingBuffer::size() will overflow) that can only be
handled there and there's no point trying to work around them here.

Reviewed-by: Markus Goetz