Fixed querying of GLX extensions under X11.
We always queried the client for its GLX extensions and used them
blindly, even though they might not be available in the server.
Also fixed the extension string check itself. We used simple sub-string
search, which could potentially break if one extension name is a
sub-string of another.

Task-number: QTBUG-5732
Reviewed-by: Kim