LineEdit graphics are sometimes missing from webpages with s60style
If the webpage does define BODY colors for background, it is possible
that lineEdits are not drawn at all. Actually, they are drawn but with
transparent background, which makes them near-invisible. This is due
to that QS60Style tries to check if the themePalette has been changed
and if it has forwards the drawing of that particular widget to common
style. Common style unfortunately uses active style's palette to draw
the widget. S60Style defines Base to be transparent (to get the
theme background shine through) and common style tries to use Base
to draw lineEdit. End result is transparent lineEdit on webpage.

This fix makes more educated deduction when it tries to see if style
has been changed. This is related to QTBUG-4885.

Task-number: QTBUG-4885
Reviewed-by: Janne Koskinen
(cherry picked from commit 75ebdd04f41dd75be1e046a1c54ced0bfb78a0f2)