Fix pixel alignment of glyphs when using raster engine on Mac Cocoa
The left bearing of the glyph has to be rounded before truncated,
otherwise the subpixel positioning of the glyph will be lost. The
visible consequence of this was on Mac Cocoa where certain pairs of
glyphs would get an extra pixel of space between them when running with
graphics system raster. With small fonts it would look pretty bad.

Task-number: QTBUG-5430
Reviewed-by: Trond