Fixed the closing of the A attribute at the start of each class page
It needs to be closed with </a> instead of the short /> notation, in
order for browsers like Firefox and Chrome not to get confused.

It caused the span used to make the subtitle smaller to only have effect
up until the link to the module started (only applying the span to the
opening '[').

Reviewed-by: Martin Smith <msmith@trolltech.com>