QListView: fixes skipping one item in pageDown(Up)
The previous implementation was scrolling down (number of visible items
in viewport) + 1 and this was leading to one item being skipped while doing
page down (the same also happened for page up). Now we are scrolling (number
of visible items in viewport) - 1 for each page down(up) (the '-1' is for
keeping the context, so the last item will turn into the first in case of
a page down, for example).

Task-number: QTBUG-5877
Reviewed-by: thierry