Fixes in the multitouch handling.
When filling in touch point data for delivering to widget don't use the setter
methods to avoid unnecessary detaches. Making it not detach also allows not to
re-calculate widget-relative start position and last position on each delivery
step.

Also use the QWeakPointer to store a pointer to a widget that expects to
receive a touch event, so that if the widget is destroyed we'll get notified.

Reviewed-by: Bradley T. Hughes