Disable linker warning 6780 in RVCT builds for Symbian OS
The linker warning that indicates symbol visibility changes is not useful
as it is an expected change to symbol visibility and produces lots of
clutter in the build logs.

(It is not desirable to follow the ELF spec when it comes to symbol
visibility in this case, which is why the linker warns and we ignore it)

This is likely to be a Raptor-only issue - I believe abld suppresses the
warning by default (at least in ABIv1 mode)

Also update gui.pro to use LFLAGS rather than MMP_RULES to alter the
arguments to ARMCC toolchain.  qmake ought to detect this (incorrect)
usage of MMP_RULES and abort MMP file generation - raised QTBU-5961 to
look at this.

Reviewed-by: Shane Kearns