Mac: Fixes broken window decorations for QGraphicsProxyWidget.
The problem was the QMacStyle didn't handle the pixel metric for
PM_TitleBarHeight correctly when passing a style option containing
an unitialized rect (empty rect).  It already had special logic for
invalid rects, so the fix is simply to extend the check to also handle
empty rects.

Auto-test included.

Task-number: QTBUG-4160
Reviewed-by: msorvig