Unified ICON keyword handling for ABLD and SBSv2 toolchains.
With SBSv2 toolchain only first file from ICON keyword is used and
it is actually the right behaviour. Originally support for multiple
files in ICON keyword was introduced to support localized application
icons. However, in that case the different localized icons should not
go to same MIF file since it is not possible to define in .rss file
which icon number should be used for which language.

To support a localized application icons separate MIF for each locale
should be created and installation of ICON should depend on selected
language. Implementing proper support for localized icons is left
for another task.

Task-number: QTBUG-4745
Reviewed-by: Miikka Heikkinen