Fix the detection of when this process gets names on the bus.
Previously, we were relying on NameOwnerChanged to notify us of our
own names. This worked because we got all NameOwnerChanged that
happened on the bus.

Now, we only get those we're interested in. Instead of watching for
newOwner==baseService, let's just use the NameAcquired and NameLost
signals, that the D-Bus server sends to us anyway.

Task-number: QTBUG-5979