configure.exe: fix when configured with -openssl-linked
before, when configured with -openssl-linked, configure.exe hang,
because of the following:
We were checking for XQMAKESPEC via something like
'dictionary["XQMAKESPEC"] == "someValue"' , which added the XQMAKESPEC
value to the dictionary. Later (line 3030), we checked whether
XQMAKESPEC was set (which was now true), and we used the value although
it was empty.

Reviewed-by: Thiago Macieira
Reviewed-by: Denis Dzyubenko
(cherry picked from commit 009f01107fd062c1cd478fc9dbf20d927ff9c459)

Conflicts:

	configure.exe