Properly return the absolute file path for help files. This fixes a bug that can only be seen on windows when we have the collection storage location on e.g. Drive c:\ while the help files reside on a different drive.
QHelpEngineCore::registerDocumentation would fail to gather the proper
relative file path, thus register the file with it's absolute path. In
QHelpEngineCore::documentationFileName we still assumed that all files
in the collection are registered relative and fail to concatenate the
proper path. Now we check if the path is already absolute and return.

Reviewed-by: ck