Revert check for file size to a warning in QFile::map
The check, introduced in c66de8a19b558e14097c402480eeabf0b5ec2351,
changes behavior and it is not clear whether we should prevent
mmaping beyond EOF. FWIW, python prevents it.

Leaving a warning in there for now so as not to break applications that
depend on the feature.

Task-number: QTBUG-6008
Reviewed-by: Thiago Macieira