Fix for compile error in Phonon MMF backend
This was introduced by d0b0b525de.

The use of QT_NO_DEBUG macros causes release-armv5 builds to fail
when the -release option is passed to configure.

This change causes the debugging code (ObjectDump and related
classes) to be included in release builds, although these functions
are not executed.

This is a temporary fix; task QTBUG-6012 has been created for
re-removing this code from release builds.

Reviewed-by: Frans Englich