Fixed antialiasing in the GL based QGraphicsView demos under X11.
All the examples/demos that used a QGraphicsView with a QGLWidget
as a viewport, that had multisampling enabled, was broken after
7b61fbf03e170a7da37d5f57ed4053aae719ec7f. This is because the reparented
QGLWidget will be forced to use the parent widget's visual, which in this
case does not contain any multisample buffers. The QX11Info structure
for QGLWidget is put together based on the QGLFormat that is passed
into the QGLWidget constructor and will in most cases always be
different from the parent widget visual.

Task-number: QTBUG-5998 and QTBUG-6003
Reviewed-by: Eskil