Fix regression introduced in 1e6b424b692b20dcfec920f8d3563e520ec1ff05
When processing the result of QIODevice::readLine, forgot to take into
account that a '\0' is appended to the array. The terminating character
is not accounted for in the number of bytes returned.

By pre-allocating a byte for the terminating null character, we make
sure we'll actually read 16k bytes on each and every iteration.

Task-number: QTBUG-6019
Reviewed-by: Thiago Macieira
(cherry picked from commit b7692016f282251002b3e85dfcb5567bd91a12c0)