Cocoa: apps with dialogs hangs on exit
This happends because we refuse to exit if we have a modal
dialog showing on screen. After some discussion, we decided
that we need to allow to exit if the quit menu item is enabled. Some
of the reason behind this is the way developers tend to (mis)use
dialogs as normal windows in Qt. So, if you don't want your app to
exit when showing a modal dialog, disable quit menuitem, or handle
QCloseEvent.

Task-number: 5613
Reviewed-by: MortenS