Mac: Make gestures available when building Qt against SDK < 10.6
We build the release packages for Carbon against the 10.4 SDK.
This means that the constants used for dealing with gestures are
missing, and currently the implementation will be ifdeffed out.
This patch defines these constants when building with an
SDK < 10.6. Gestures will therefore work when running applications
on 10.6. The same fix is also implemented for cocoa.

Rev-By: MortenS