Fix positioning of diacritics in .otf fonts on Windows
Since .otf fonts are not considered truetype fonts, they take the
code paths intended for non-outline fonts. When calculating the bounding
box of a glyph, this would mean we'd find the largest possible bounding
rect of any glyph in the font, while the other metrics, such as the GPOS
tables used to position diacritics in relation to base glyphs, are
positioning the actual outline of the glyph. The result was that certain
diacritics that depended on the opentype positioning would not be shown on
Windows at all, as they would be positioned based on the wrong left bearing
and height when drawn into the glyph cache.

The fix is to find the tight bounding rect of the outline whenever
possible and fall back to the old code when this fails. I've also added
the left bearing of the glyph to the bounding box in the fallback case,
as we did not respect this before and would misplace glyphs that has
a bearing.

Task-number: QTBUG-5860
Reviewed-by: Trond