QNetworkCookieJar: don't do path checking when receiving cookies
Actually, the RFC 2109 says cookies should not be stored if the path
attribute is not a prefix of the request URI the cookie comes from.
However, all browsers allow it anyway; with the demo browser e.g.
logging in to wordpress.com was not possible.
We still do path checking when sending cookies, i.e. in
QNetworkCookieJar::cookiesForUrl().

Reviewed-by: Markus Goetz
Task-number: QTBUG-5815