Improved and optimized the WaveUnderline
The WaveUnderline used to be a series of 180 degree arcs, that didn't
always align very well between different text items and that could be
clipped at the bottom since the font height got reduced. Now it uses
quad beziers so that it is more like a sinus curve, and the alignment
and clipping issues have been fixed.

The painting of the WaveUnderline has been optimized using QPixmapCache.
A 100 pixels wide sample of the wave is now cached as a pixmap and
repeated to draw the WaveUnderline. This is an order of magnitude faster
than rasterizing a QPainterPath.

The QPlainTextEdit needs to set the brush origin since it draws the text
items in local widget coordinates, rather than relying on a painter
translation like the QTextEdit.

Done with mae.