Fix softkeys for QDialogButtonBoxes created without a parent.
Inthe case where a QDialogButtonBox was created without a parent, for
example in the FTP example, the softkeys that are automatically created
inside the button box were not being added to the right widget when the
button box was given a parent (or added to a layout) later.

This patch resolves that issue by handling the ParentChange event and
then adding the softkeys at this point.

Task-number: QTBUG-6086
Reviewed-by: axis