Fix crash in QApplication::beep() on S60
Using an asynchronous API synchronously caused a crash, because the
data structure was deleted while an async request is still pending.
Depending on the multimedia implementation on the phone, the crash will
happen or you just get no sound (or it can even work if the underlying
implementation is blocking).

Solution is to use the API asynchronously, and delete the handling object
in qt_cleanup.

Also raised the tone by one octave to be more similar to the system beep.

Task-number: QTBUG-6170
Reviewed-by: Janne Anttila
Reviewed-by: Miikka Heikkinen