Bind the EGL API type before calling eglMakeCurrent()
When mixing GL and VG in the same thread, some devices require
that eglBindAPI() be called before eglMakeCurrent() when switching
between context types.  The EGL spec indicates that this shouldn't
be required unless the context parameter is EGL_NO_CONTEXT, but it
isn't a big deal to work around slightly non-compliant devices.

Reviewed-by: Tom Cooksey