Make QVGCompositionHelper::blitWindow() more generic
Previously, the composition helper would use the VG window surface
private structure to get the VGImage to blit.  This change passes
the VGImage and size values in directly so that the composition
helper can also be used to blit GL surfaces that have been converted
into a VGImage via an EGLImage.

Reviewed-by: Tom Cooksey